Rename root package to backdrop
