+++ backdrop/write/api.py
@@ -11,7 +11,7 @@ app = Flask(__name__)
 
 # Configuration
 app.config.from_object(
-    "performance_platform.write.config.%s" % getenv("GOVUK_ENV", "development")
+    "backdrop.write.config.%s" % getenv("GOVUK_ENV", "development")
 )
 
 mongo = MongoClient(app.config['MONGO_HOST'], app.config['MONGO_PORT'])
