+++ backdrop/write/config/test.py
@@ -1,3 +1,3 @@
-DATABASE_NAME = "performance_platform"
+DATABASE_NAME = "backdrop_test"
 MONGO_HOST = 'localhost'
 MONGO_PORT = 27017
