+++ collectors/shove-data-into-mongo.sh
@@ -1,1 +1,1 @@
-python licensing_location_data.py | mongoimport --db performance_platform --collection licensing --upsertFields licence,authority,interaction,location,start_at,end_at --jsonArray
+python licensing_location_data.py | mongoimport --db backdrop --collection licensing --upsertFields licence,authority,interaction,location,start_at,end_at --jsonArray
