+++ tests/write/unit_tests.py
@@ -4,8 +4,8 @@ from hamcrest import *
 
 import pytz
 
-from performance_platform.write import api
-from performance_platform.core.validators import value_is_valid_id, \
+from backdrop.write import api
+from backdrop.core.validators import value_is_valid_id, \
     value_is_valid, key_is_valid, value_is_valid_datetime_string
